+++ examples/echoserver.php
@@ -11,7 +11,7 @@
  * Console options:
  *  --port <int> : The port to listen to, default 80
  *  --ssl : Use SSL, default false
- *  --timeout <int> : Timeout in seconds
+ *  --timeout <int|float> : Timeout in seconds
  *  --framesize <int> : Frame payload size in bytes
  *  --connections <int> : Max number of connections, default unlimited
  *  --deflate : Add support for per-message deflate compression
@@ -48,7 +48,7 @@ echo "# Echo server! [phrity/websocket]\n";
  * @var array{
  *     port: int<1, 32768>,
  *     ssl: bool,
- *     timeout: int<0, max>,
+ *     timeout: int<0, max>|float,
  *     framesize: int<1, max>,
  *     connections: int<0, max>|null,
  *     deflate: bool,