+++ docs/Changelog.md
@@ -8,6 +8,8 @@
 
 ### `3.5.0`
 
+ * Timeouts support float (@sirn-se)
+ * Client and Server start() can have specific timeout (@sirn-se)
  * Internal LoggerAwareTrait for logger handling (@sirn-se)
  * Constant in interface (@sirn-se)
  * Frame pulling rewritten (@sirn-se)